# Optimize language server performance