--- a/sway-lsp/src/core/token.rs
+++ b/sway-lsp/src/core/token.rs
@@ -259,3 +259,19 @@ pub fn get_range_from_span(span: &Span) -> Range {
         end: Position::new(end_line, end_character),
     }
 }
+
+pub fn get_range_from_span_opt(span: &Span) -> Range {
+    let start = span.start_pos_opt().line_col();
+    let end = span.end_pos_opt().line_col();
+
+    // let start_line = start.0 as u32 - 1;
+    // let start_character = start.1 as u32 - 1;
+
+    // let end_line = end.0 as u32 - 1;
+    // let end_character = end.1 as u32 - 1;
+
+    Range {
+        start: Position::new(start.0 as u32 - 1, start.1 as u32 - 1),
+        end: Position::new(end.0 as u32 - 1, end.1 as u32 - 1),
+    }
+}