… sway-lsp/src/core/token_map.rs
… sway-lsp/src/core/token_map_ext.rs
… sway-lsp/src/server_state.rs
… sway-lsp/src/traverse/dependency.rs
… sway-lsp/src/traverse/lexed_tree.rs
… sway-lsp/src/traverse/mod.rs
… sway-lsp/src/traverse/parsed_tree.rs
… sway-lsp/src/traverse/typed_tree.rs
… sway-lsp/src/utils/debug.rs
… sway-types/src/span.rs