@xunilrj 2023-08-16
>  Removing the clones led to a 40% speed up.

Brilliant!

> The line_col() method was getting called a lot in order to convert between the span's start and end to the lsp_types::Range type that is used in LSP. As such, we were calculating this for every token in the language server when iterating over the TokenMap for all of the features.

I am also having this problem elsewhere. Maybe we should have a data struct `offset -> line number` somewhere? `crates.io` has some interval trees. Not sure which one would be the best fit.