@JoshuaBatty 2023-08-18
> What do you think about the `Position` type taking a `SourceId` instead? Would that provide any benefits or just complicate things for you?

I'm not sure I understand the proposed benefit of doing that. One of the goals of this work is to get away from the `Position` and `SourceId` types as quickly as possible. This way we only need to compute the `Range` once. We also eliminate the need to pass around the source engine throughout the server to get the required `PathBuf`.

Would be keen to hear your thoughts on if you see a better way forward though!