@JoshuaBatty 2023-08-18
> Maybe we should have a data struct `offset -> line number` somewhere? `crates.io` has some interval trees. Not sure which one would be the best fit.

Yeah that could be interesting! Ping me if you come across anything worth looking into